packages/mory/mory-0.5.0.tar.gz/mory-0.5.0/tests/test_main.py
import pytest
from unittest.mock import patch, MagicMock
import argparse
from mory.main import parse_args, get_root_folder

@patch('mory.main.prompt')
def test_get_root_folder(mock_prompt):
    mock_prompt.return_value = '/fakepath'
    path = get_root_folder(None)
    assert path == '/fakepath'

    path = get_root_folder('/anotherpath')
    assert path == '/anotherpath'

@patch('argparse.ArgumentParser.parse_args')
def test_parse_args(mock_parse_args):
    mock_parse_args.return_value = argparse.Namespace(command='scan', path='/fakepath', save_to_file=False, structure_only=False, git_skip=False)
    args = parse_args()
    assert args.command == 'scan'
    assert args.path == '/fakepath'
    assert not args.save_to_file
    assert not args.structure_only

# Additional test cases can be added for other command handlers and edge cases
